change type of member type of Task
type is now TaskType instead of uint32_t

fixed compile errors from this change in:

- Task (obviously, also adjusted serialize and deserialize)
- Client, Scheduler, Worker as well as their CLP's and CLP's tests respectively
- All of the ClientRequest Tests
- ConnectTest

+++ src/scheduler/Scheduler.cpp
@@ -35,7 +35,7 @@ void Scheduler::processCommandLineArguments(int argc, const char* const * argv)
 {
     SchedulerCommandLineProcessor clp;
     clp.process(argc, argv, task);
-    if(task->getType()) {
+    if((uint32_t)task->getType()) {
         std::string server = "localhost";
         short port = 8443;
         if(!task->getServerIP().empty()) {
