+++ testsrc/scheduler/clientRequests/TestContinueRequest.cpp
@@ -22,7 +22,7 @@ TEST(TestContinueRequest, allArgs)
     auto task = std::make_shared<Task>();
     auto config = task->getConfig();
 
-    task->setType((uint32_t)TaskType::CONTINUE);
+    task->setType(TaskType::CONTINUE);
     task->setJobId(0);
 
     auto response = req->executeRequestAndFetchData(task, dbGetJob, dbUpdateJobStatus, dbAddJob, userID);
@@ -36,7 +36,7 @@ TEST(TestContinueRequest, noArgs)
     auto task = std::make_shared<Task>();
     auto config = task->getConfig();
 
-    task->setType((uint32_t)TaskType::CONTINUE);
+    task->setType(TaskType::CONTINUE);
     task->setJobId(std::nullopt);
 
     auto response = req->executeRequestAndFetchData(task, dbGetJob, dbUpdateJobStatus, dbAddJob, userID);
